Workaround incorrect SCRIPTS by using PREFIX
The `SCRIPTS` environment variable points to the old host environment
used for testing. However this is incorrect during the test phase. To
fix this issue, switch to using the `PREFIX` environment variable, which
is set correctly, and add the appropriate path after to point to where
`SCRIPTS` should have pointed.