Update `DTSDKName` to match our macOS SDK
The `DTSDKName` was getting fixed by our old `sed` strategy along with
the `MinimumSDKVersion`. We fix the `MinimumSDKVersion` already as this
vital. It's less clear if the `DTSDKName` needs to be accurately
reflected. Though it could show up in applications' `Info.plist`, so we
might as well have it accurately reflected just in case. After this
change there does not appear to be any more references to the OS SDK
version other than the intended one.