Don't fail run_conda_forge_build_setup when conda-forge.yml missing.
The run_conda_forge_build_setup scripts read from conda-forge.yml to get
the channel_priority, defaulting to strict if the setting is missing.
But if conda-forge.yml was missing completely, the script would fail
because shyaml does not return the default when there is no YAML to
parse at all. This fixes that by ensuring that "strict" is returned in
those cases as well.