Drop outdated fixme comment
This comment may have made sense before CUDA 12 was added (which has all
of the packages this comment mentions) or before CUDA 11.8 was added
(which follows the same build model as 11.2). However now that both of
those are in and it unlikely that other CUDA 11.x's would be added,
think it is safe to say this work is no longer needed and won't be done.
Maintainers and users that want these packages can move to CUDA 12.