Add `targets` content from `cuda-nvcc` for CUDA 11
Within `cuda-nvcc`'s RPM's `targets` directory, there is some content
needed for `nvPTXCompiler`, which the Numba supporting CUDA 11 package,
`ptxcompiler`, uses. This makes sure it is available when
cross-compiling `ptxcompiler`.

As the Docker image and `conda-forge-ci-setup` limit content to the
`targets` directory, the rest of the `cuda-nvcc` content will be
ignored. This is for the best to ensure these other bits are not pulled
in accidentally.